made necessary changes for merging of the two split functions. Minor changes in home.jsx
