+++ frontend/src/components/split.jsx
