--- frontend/src/components/ts.jsx
@@ -1,31 +1,0 @@
-import React from 'react'
-
-export default () => {
-	return (
-		<div>
-			{/* Page Heading */}
-			{/*<div className="d-sm-flex align-items-center justify-content-between mb-4">
-				<h1 className="h3 mb-0 text-gray-800"><br />GTFS-ride Info</h1>
-			</div>*/}
-			
-			{/* Content Row */}
-			<div className="row">
-
-				{/* Content Column */}
-				<div className="col-lg-6 mb-4">
-
-					{/* Project Card Example */}
-					<div className="card shadow mb-4">
-						<div className="card-header py-3">
-							<h6 className="m-0 font-weight-bold text-primary">Under construction</h6>
-						</div>
-						<div className="card-body">
-							Coming soon ...
-						</div>
-					</div>
-				</div>
-			</div>
-	</div>
-
-	)
-}
